+++ NEWS.md
@@ -1,5 +1,6 @@
 - added initial support for js functions with `js_function()`
 - fix return type for ts_undefined(), ts_void(), ts_union() (#6)
+- add unnamed array-type list
 
 # ts 0.2.0